:white_check_mark: test: split into multiple testcode modules
